+++ builder_pkgs/mockito/CHANGELOG.md
@@ -3,6 +3,9 @@
 * Ignore "must_be_immutable" warning in generated files. Mocks cannot be made
   immutable anyway, but this way users aren't prevented from using generated
   mocks altogether.
+* Require Dart >= 3.3.0.
+* Require analyzer 6.4.1.
+* Add support for extension types.
 
 ## 5.4.4